+++ src/include/string_work.h
@@ -1,8 +1,7 @@
 #ifndef ONEGIN_HEADER
 #define ONEGIN_HEADER
 
-enum ONEG_ERRS
-{
+enum ONEG_ERRS {
   ARGC_ERR = 1111,
   OUTPUT_OPN_ERR,
 };
@@ -11,6 +10,6 @@ int str_cmp(const void *ptr1, const void *ptr2);
 
 int reverse_str_cmp(const void *ptr1, const void *ptr2);
 
-void string_print(FILE *out, string *strings, int str_number);
+void string_print(FILE *out, string *strings, size_t str_number);
 
 #endif
